backend/requirements.txt
django>=4.2
djangorestframework>=3.14
reportlab>=4.0
Pillow>=10.0
PyPDF2>=3.0
python-dotenv>=1.0
